+++ CHANGES.rst
@@ -5,7 +5,8 @@ Change log
 0.2 (unreleased)
 ================
 
-- Nothing changed yet.
+- Add ``--start`` option to the migrate script to be able to start analysis
+  with a predefined OID.
 
 
 0.1 (2016-05-19)